downscope: move example files to a separate file & package
